move EntityWithNullForMissingField
